adjusted wagmi generate script to regenerate hooks
